- put the credentials on env vars - get mime type and file size on upload

+++ models/media.go
@@ -8,7 +8,7 @@ type Media struct {
 	gorm.Model
 	FileName string `gorm:"size:255;not null" json:"file_name"`
 	FileSrc  string `gorm:"size:255;not null" json:"file_src"`
-	FileSize uint   `gorm:"default:0" json:"file_size"`
+	FileSize int64  `gorm:"default:0" json:"file_size"`
 	MimeType string `gorm:"size:30;" json:"mime_type"`
 	Title    string `gorm:"size:255;not null" json:"title"`
 	AltText  string `gorm:"size:255" json:"alt_text"`
